Modified assignment release date implementation.
